Refactor number request handlers
